refactored server skeleton and added shared directory to development environment
