+++ src/server_main.cc
@@ -1,63 +1,10 @@
 #include <boost/asio.hpp>
 #include <iostream>
 
+#include "http_server.h"
+
 using namespace boost::asio;
 using namespace boost::asio::ip;
-
-class http_session : public std::enable_shared_from_this<http_session> {
-public:
-  http_session(tcp::socket socket) : socket_(std::move(socket)) {}
-
-  void start() {
-    do_read();
-  }
-
-private:
-  void do_read() {
-    auto self(shared_from_this());
-    socket_.async_read_some(buffer(data_, max_length),
-        [this, self](boost::system::error_code ec, std::size_t length) {
-          if (!ec) {
-            do_write(length);
-          }
-        });
-  }
-
-  void do_write(std::size_t length) {
-    auto self(shared_from_this());
-    boost::asio::async_write(socket_, boost::asio::buffer(data_, length),
-        [this, self](boost::system::error_code ec, std::size_t /*length*/) {
-          if (!ec) {
-            do_read();
-          }
-        });
-  }
-
-  tcp::socket socket_;
-  enum { max_length = 1024 };
-  char data_[max_length];
-};
-
-class http_server {
-public:
-  http_server(boost::asio::io_context& io_context, short port) :
-      acceptor_(io_context, tcp::endpoint(tcp::v4(), port)) {
-    do_accept();
-  }
-
-private:
-  void do_accept() {
-    acceptor_.async_accept(
-        [this](boost::system::error_code ec, tcp::socket socket) {
-          if (!ec) {
-            std::make_shared<http_session>(std::move(socket))->start();
-          }
-          do_accept();
-        });
-  }
-
-  tcp::acceptor acceptor_;
-};
 
 int main(int argc, char* argv[]) {
   try {
@@ -67,7 +14,6 @@ int main(int argc, char* argv[]) {
     }
 
     boost::asio::io_context io_context;
-
     http_server server(io_context, std::atoi(argv[1]));
 
     io_context.run();
